Fixed mass regression.
Fixed an earlier bad commit caused by trusting my code editor 'find usages' for several functions which led to the deletion of code that was actually being used.
'$this' passed as a variable works, instead of having extra '.'.
Arrays are now processed properly in a stageful way, rather than as a regex. So they not only work at all, but they also work recursively.