Added support for try/catch, with $e.getMessage() being converted to e.message. Added support for type-hinting.
Both of these are reasonably fragile and have only been tested gently.

E.g. for type-hinting, in the code

function testFunction(array $testArray){
…
}

The first array actually generates a T_ARRAY token, which has to be suppressed just because we are in a CatchScope. If other tokens, other than just strings can be passed in, then they will need to similarly be suppressed.